+++ src/functions.py
@@ -4,79 +4,124 @@
 
 # Import statements
 import xml.etree.ElementTree as ET
-
+import re
+import os
+import csv
+from sklearn.feature_extraction.text import TfidfVectorizer 
+from bs4 import BeautifulSoup
 
 # Functions
-def extract_text(xml_file) -> str:
+def extract_text(file_path: str) -> str:
     """
-        This function takes in an earnings transcript as an input, and extracts the words from the transcript.
 
-        Input: .xml file
-        Output: str
+    This function takes in an earnings transcript as an input, and extracts the words from the transcript.
+
+    Input: .xml file
+    Output: str
     """
-    try:
-        tree = ET.parse(xml_file)
-        root = tree.getroot()
-
-        # find <body> tag and extract
-        body = root.find(".//Body")
+    
+    # Parse the XML file
+    tree = ET.parse(file_path)
+    root = tree.getroot()
+    
+    # Extract all text content from XML elements, including CDATA
+    text_parts = []
+    
+    # Look specifically for Body and Headline elements which contain CDATA
+    for event_story in root.findall('.//EventStory'):
+        # Extract headline
+        headline = event_story.find('Headline')
+        if headline is not None and headline.text:
+            text_parts.append(headline.text.strip())
+        
+        # Extract body
+        body = event_story.find('Body')
         if body is not None and body.text:
-            # clean and return
-            return body.text.strip()
-        else:
-            return "No body section found."
+            text_parts.append(body.text.strip())
+    
+    # Join all text parts with spaces
+    raw_text = ' '.join(text_parts)
+    
+    # Clean up the text
+    # Remove multiple spaces
+    cleaned_text = re.sub(r'\s+', ' ', raw_text)
+    # Remove special characters and normalize whitespace
+    cleaned_text = re.sub(r'[^\w\s.,?!-]', '', cleaned_text)
+    # Trim leading/trailing whitespace
+    cleaned_text = cleaned_text.strip()
+    
+    return cleaned_text
 
-    except ET.ParseError as e:
-        return f"Error parsing XML: {e}"
-    except Exception as e:
-        return f"An error occurred: {e}"
 
-# test
-# spoken_words = extract_text("src/data/example.xml")
-# print(spoken_words)
-
-def extract_exposure(exposure_words, txt_string, buffer) -> dict:
+def csv_to_list(filepath):
     """
-    This takes in the str returned from extract_text, and extracts regions (+- buffer) where the exposure
-    words exist.
-
-    For example, if buffer = 5, then wherever we identify an exposure word, we take the substring of words beginning
-    5 words before the exposure word, and 5 words after the exposure word. This would create a string with 11 words.
-    We would then add this to our return dict.
-
-    Input: 
-    - exposure_words: csv string of words (e.g., "smoke, pollution, radiation")
-    - txt_string: str containing the text to search in
-    - buffer: int number of words before and after the exposure word to include in the context
-
-    Output:
-    dict containing extracted context strings around each occurrence of an exposure word.
-    The keys are generated identifiers (e.g., "exposure_0", "exposure_1", etc.)
+    This takes in a CSV file consisting of political bigrams and returns
+    a list of bigrams with the underscore removed (e.g., "democratic_party" -> "democratic party").
     """
+    political_list = []
     
-    # split the exposure_words CSV string into a list
-    exposure_list = [word.strip().lower() for word in exposure_words.split(',') if word.strip()]
+    with open(filepath, 'r', encoding='utf-8') as file:
+        reader = csv.reader(file)
+        
+        # Convert the CSV rows to a list so we can manipulate them outside the 'with' block
+        rows = list(reader)
     
-    words = txt_string.split()
+    # Remove the first row if it's a header
+    if rows:
+        rows.pop(0)
     
-    exposure_contexts = {}
-    context_count = 0
+    # Process each row, replacing underscores with spaces
+    for row in rows:
+        # Ensure row is not empty
+        if row and len(row) > 0:
+            bigram = row[0]
+            # Replace the first underscore with a space if you only expect two parts
+            parts = bigram.split("_")
+            if len(parts) == 2:
+                bigram = parts[0] + " " + parts[1]
+            political_list.append(bigram.lower())
     
-    for i, word in enumerate(words):
-        if word.lower() in exposure_list:
-            start = max(0, i - buffer)
-            end = min(len(words), i + buffer + 1) 
+    return political_list
+
+
+def extract_exposure(text, keywords, window=10) -> dict :
+    """
+        This takes in the str returned from extract_text, and extracts regions (+- buffer) where the exposure
+        words exist. 
+
+        For example, if buffer = 5, then wherever we identify an exposure word, we take the substring of words beginning
+        5 words before exposure word, and 5 words after the exposure words. This would create a string with 11 words. 
+        We would then add this to our return dict.
+
+        Input: 
+        - exposure_words: csv
+        - txt_string: str
+        - buffer: int
+
+        Output:
+        dictionary
+        
+    """
+    words = re.findall(r'\w+', text)
+    contexts = {}
+
+    for index, word in enumerate(words):
+        if word.lower() in keywords:
+            start = max(0, index - window)
+            end = min(len(words), index + window + 1)
             context = " ".join(words[start:end])
-            exposure_contexts[f"exposure_{context_count}"] = context
-            context_count += 1
+            contexts[word] = context
 
-    return exposure_contexts
+    return contexts
 
 
 def sentiment_score(text_dict):
     """
     Returns sentiment scores for each string in text_dict using RoBERTa-based
     sentiment analysis for positive/negative/neutral sentiment.
+
+    TODO:
+    - reference to how the sentiment reference works
     """
     from transformers import pipeline
     
@@ -109,7 +154,22 @@ def sentiment_score(text_dict):
 
     return results
 
+def tf_idf(*args):
+    '''
+        Input: 
+        Output:
+    '''
+    documents = list(args)
+    if not all(isinstance(doc, str) for doc in documents):
+        raise ValueError("All inputs must be strings.")
 
+    tfidf = TfidfVectorizer()
+    result = tfidf.fit_transform(documents)
 
+    print('\nIDF values:')
+    for word, idf in zip(tfidf.get_feature_names_out(), tfidf.idf_):
+        print(word, ':', idf)
+
+    return result
